Fixed Guest Access bug + Removed large Keyboard Physics + Sped up sphere selection
Fixed bug with Guest access - switching in and out wouldn’t refresh the state you were in, hence it would continue to try and view images you may not have access too.

Removed the limitation of only using the keyboard whilst its up – I’d added this limitation because if you select a button while the keyboard is up then you realistically need to AcceptText() in the keyboard first, and so I had to go around adding this call in a bunch of places!

Some images were still being uploaded quicker than in 3 seconds, so I made it such that the ImageSphere coming towards takes only 2 seconds!