Updated MaxWidth and ThumbnailWidth
I've introduced a MaxWidth for images because to be fair, I think its pretty illogical to allow users to upload images at whatever resolution they want to... what if someone uploaded a ginormous image, it could fuck up the system somewhere along the way...

Therefore I've now added an upper bound to images being uploaded to the server, which I've currently set to a 4k width (2^12 = 4096). Whenever I upload to the server I create both the Thumbnail (as I did before) and this new MaxResolution image if the image is not below that resolution already. I think this is an entirely reasonable size and it will give me a much greater guarantee of the speed images will load at.

The only issue I now have is that Gallery images still load at a huge size. But I think with the new OpenGL code that Andy is giving me, we may be able to fix this!