Got the app ready for V2 submission!
Made the usual necessary updates of setting it into (1) Production mode, (2) debug off, and (3) AndroidManifest has INFO set instead of LAUNCHER.

Also changed and fixed numerous things...
(1) Gallery images and Profile spheres also open new pages in the ImageFlow, now allowing the user to activate the MenuHider and see the entire image without the menu or the dim-effect.
(2) Since we have the ImageFlow for the Gallery, I've stuck "Create Post" and all its UI within ImageFlow as opposed to within Gallery, which feels a lot cleaner and has also removed the ImageSkybox's job of looking after the UploadButton, which was never quite right.
(3) Fixed the CommentIcon selected colour in the ImageSummary, as well as the comment count itself.
(4) Updated Oculus Utilities to the latest possible.

The C++ Plugin has Andy Thomason's new resampling code, although its not currently active.