Added Thumbnails to Gallery and Upper limit to ImageSize
This is essentially done in the C++ Plugin through Andy Thomason's code. The key thing is that whenever loading an image, I tell it the maximum size it can be. Therefore, I can set the maximum to 512, or set the maximum to 4096, and hence simulate Thumbnail generation simply by loading the images at a lower resolution =D

There appears to be a bug however that manifests itself when loading thumbnails for the bigger pictures that the Samsung Gear360 takes (it doesn't appear with the Insta360 images), and they end up looking grey and choppy like they would in a television screen that's got poor signal.

Even with this bug I still think its worth checking in because this is a substantial improvement to Gallery Image loading and I've let Andy know about my bug so that he can come up with a fix!

In the process I also fixed an annoying bug with Images loading too fast and being splatted over by their Thumbnail. The way to fix this was simply to make sure that I check whether the ImageLoader has finished its Loading... if so then I don't splat the Thumbnail onto the background because its already happened!