Created option to view images in RGB565 + Partially added Keyboard TextField
Added Andy Thomason's code to resample images either in RGB565 (16-bit) or the standard RGB (24-bit). I've added a flag that allows me to toggle between them so that at some point if we think the quality has dropped then we can check and switch back to 24-bit.

Through my testing it seems to not have changed image quality at all though and switiching over to RGB565 should be a significant saving in principle (although I haven't yet checked and will definitely do so) so I've set that to be true as our default!

With Mox I also re-positioned the Keyboard, shrank it, and created a TextField that I need to get working properly.