Keyboard TypeBox works
The keyboard is smaller but and at the same Z-depth as the BackArrow, all BackArrows will call CancelText() just in case its open, and the TypeBox now works too!

I spotted some weird shearing going on with the VR effect when tilting the head left and right, which seems to happen just as much with the latest version of the app as with this build, which upstes me a lot =( - I'm going to take a look at this next.