VR Shear bug Temporary fix + Deleted Extensions folder
One of the most unusual bugs I've ever encountered has cropped up and I'm submitting a temporary fix for it.

Essentially, since the latest submission of new assets (screenshots and the like for the Oculus store) went through on Friday 31st, a shearing effect has cropped up in the app that I managed to track down to its occurence being simply tied to the name of the package!

So when I name any project "io.vreel.vreel" (even an entirely new one in Unity) and build the app, then this shearing effect happens! Otherwise, if I give the app any other package name, such as "io.vreel.vreel_noshearing" (as I have) then the shearing won't occur!

Now obviously this isn't a long-term solution because for the app to be uploaded to the store it needs to have the same package name as the previously uploaded one. Moreover, this feels like a bug somewhere in the Oculus software interface to the GearVR, and is something that I certainly need to fix with their help.

Either way I'm checking in this renaming of the package as a temporary fix to the problem.