Made the Keyboard more responsive!
MenuButton now has an "OnButtonDownFunc" that gets called when the HandleDown() is called.

For most buttons I'll just ignore the ability to tie something up to this function. However for the keyboard all the keys that type call this function, hence making it far more responsive as you don't need to wait for the HandleUp() to be called before the Keyboard responds to the key press!

With regards to my bizarre shearing bug I've seen that the bug is indeed entirely down to the package name I give it. Yet, a name that doesn't cause the shear eg. "io.vreel.vreel_noshear" can go onto causing the shear later on after a few builds - this has led me to feel more confused about the bug, and less sure of what the actual problem is =(