Added the ability for Sphere to Rotate
Now the user can make the ImageSphere rotate when he holds onto it and drags it. This is perfect for being able to see the entire image before even selecting it.

Next I will make the sphere rotation be the rotation that you view the image on the skybox at, when you select it!