Small fix to Animation_03
The profile name was inconsistent with the profile description in the video - as we had "david12" and "arthur", so I corrected it by making them both into "arthur".

The video looks really great on the whole though and I think once we get the animations into the app it will add a lot of value to it.