Menu is now hidden by default when Image Selected!
This was a suggestion from Simone, which makes a lot of sense. We now see the image after the sphere has been selected, rather than a comment.

I also fixed the Image selection such that when you are rotating it, it won't be accidentally selected. However, Simone actually suggested that it should be rotated by swiping the sphere, rather than using your face, which in fact makes a lot of sense, so I'm going to fix this next...