Swiping on touch pad now rotates Spheres
This is a much simpler implementation code-wise than what I had before - as before I was doing 3D math with vectors, comparing angles, and now I'm simply comparing x-coordinates along a touch-screen (ie. 1D).

However, the new method is a lot more intuitive and feels a lot better as the user, because you simply drag your finger on the touch pad and it moves the sphere!

The actual speed of the swipe and senstivity can definitely do with improving!