Updated images in VReel Video and added music
Updated the images in VReel Video and added music, and changed some global settings in order to reflect that we were building the video.

There's currently still a problem that the sphere's are turning around once they've finished their expansion, and the output video is not being created at a good quality - so I'm going to try and find a Unity plugin that spits out the video.