Added ability to rotate outside sphere when not in VR
We're now thinking of building an Android version that would go onto the Google PlayStore and switch into Cardboard mode. The idea is to keep the design the same but allow the user to spin the image around with his finger, so I've began doing this by essentially copying the sphere roation code I've stuck on all the spheres.

However I've spotted two problems:
(1) The VideoCapture script has gone back to not working and just giving me a "Camera not Rendering" message onscreen.
(2) The current version of the app that we're building seems to be either washing out all the 360 images if we're not in VR, or darkening the main sphere if we're in VR  - this does not happen in the version that's on the Oculus Store, therefore I need to work out why this is happening.