Fixed the washed out colours bug!
Somehow a couple of check-ins ago I'd set the ColourSpace in ProjectSettings to Linear instead of Gamma and that had caused the colours to be skewed in all builds!

Its annoying that I lost time trying to figure this out, but at least I now know that if I see unusual washed out colours as I did, then the problem is the Colour Space flag!