Spinning SkyboxSphere now works on X and Y axis
It took me a while to figure this out, but the key was to make the rotation along world-y and local-x!

I also managed to improve the way rotational damping occurs by storing the speed on release and reducing the currspeed by a factor of that original speed every frame!

This essentially makes the app usable in non-VR mode, except for a couple of things which still need improving: (1) the UI needs to be larger so that the keyboard is usable, and (2) the MenuHider needs to be smaller so that it doesn't block users swiping the skyboxsphere!