Made a couple of changes that make the NonVR version far better
We want to have a non-VR version of the VReel app too so I've been trying to make the app work in a pleasant way outside of VR.

In the last check-in I made it possible for the user to spin the skybox properly with their finger.

In this check-in I'm now also doing the following:
(1) Making it such that the MenuHider works differently in NonVR - as it will Hide the Menu when you touch the Skybox and Show it when you select the Logo.
(2) The AppDirector is now also responsible for changing the FOV on the camera and setting the Keyboard size appropriately if in VR or not - this is because by decreasing the FOV in NonVR we get a much better size for the UI, plus we give the Keyboard an extra boost to make it usable.