Created the first steps of the Carousel
I downloaded a Spline asset from the asset store and made the ImageSphere's follow the Spline on the Carousel when the user swipes along the scroll button I added.

This all works very nicely and simply, and now I need to get spheres that have fully rotated to load the next image in the Posts or Gallery!