Got Carousel working!
I'm sure there will be bugs around, but for the most part the Carousel now works! =D

Essentially when you hold one of the side arrows down it will now push the sphere's along the spline and whenever a sphere wraps around to the other side of the spline it (1) erases the current texture and metadata, and (2) asks Posts/Gallery to bring up the next/previous image.

This has left parts of the code a bit messy, especially Posts and Gallery who have the some functions (like NextPosts() ) still there but commented out in favour of the new function (in this case NextPost()).

I changed the scrolling mechansim to use the Arrows instead of that make-shift scrollbar that I was using, which has itself been deleted. SelectArrows are now hiding only when the Carousel has a sphere in the middle and the Posts/Gallery are at the start/end.

The most noticeable bugs I've spotted are related to menuhiding in NonVR when you select the background in some unusual situations.