Added Andy's "UseExif" code - but disabled it because I couldn't get it to work
Andy Thomason wrote two chunks of code in order to help speed up the image-loading in VReel.

(1) He wrote a way of making use of stbi_image which takes the image being downloaded and reads the data coming in from the internet, one chunk at a time, speeding up the process of downloading+decoding an image. ( https://gist.github.com/andy-thomason/8b49534aa543de1726090e33a2ac222f )

(2) He wrote a chunk of code for decoding exifs in jpgs so that I would be able to speed up the displaying of gallery images, ie. images on the phone itself. ( https://gist.github.com/andy-thomason/994db9582d816eb0c9a65006753e96f4 )

Now, I tried to bring in the second chunk of code to VReel's code-base and found that it didn't just work outright (probably because I'm doing something wrong, I doubt its related to Andy's code).

So I'm checking this in order to not lose it, but this isn't exactly a successful checkin, and the UseExif flag is always being set to false.